+++ lib/Braintree/PaymentMethodGateway.php
@@ -140,8 +140,11 @@ class PaymentMethodGateway
             'verificationAmount',
             ['paypal' => [
                 'payee_email',
+                'payeeEmail',
                 'order_id',
+                'orderId',
                 'custom_field',
+                'customField',
                 'description',
                 'amount',
             ]],